refactor: remove internal package and simplify module structure
- Delete `com.xenoterracide.git.internal` package, moving classes to `com.xenoterracide.git`.
- Update visibility of utility classes and methods.
- Rename `semver` artifact to `semver-calculator`.
