+++ module/git-metadata/src/main/java/com/xenoterracide/git/GitMetadata.java
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: Copyright © 2024-2026 Caleb Cushing
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
 
 package com.xenoterracide.git;
