+++ module/git-metadata/src/main/java/com/xenoterracide/git/TryGit.java
@@ -1,9 +1,8 @@
 // SPDX-FileCopyrightText: Copyright © 2024-2026 Caleb Cushing
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
 
-package com.xenoterracide.git.internal;
+package com.xenoterracide.git;
 
 import io.vavr.CheckedFunction1;
 import io.vavr.control.Try;
