+++ module/git-metadata/src/main/java/module-info.java
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: Copyright © 2024-2026 Caleb Cushing
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
-// SPDX-License-Identifier: MIT
 
 /**
  * Git metadata extraction module.
@@ -17,5 +16,4 @@ module com.xenoterracide.git {
   requires static org.jspecify;
 
   exports com.xenoterracide.git;
-  exports com.xenoterracide.git.internal;
 }
