+++ module/git-metadata/src/test/java/com/xenoterracide/git/DistanceCalculatorTest.java
@@ -9,7 +9,6 @@ import static com.xenoterracide.git.fixtures.CommitTools.commit;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.google.common.base.Splitter;
-import com.xenoterracide.git.internal.DistanceCalculator;
 import java.io.File;
 import org.eclipse.jgit.api.Git;
 import org.junit.jupiter.api.Test;
