--- module/git-metadata/src/test/java/com/xenoterracide/git/NoExecSystemReaderTest.java
@@ -1,27 +1,0 @@
-// SPDX-FileCopyrightText: Copyright © 2024-2026 Caleb Cushing
-//
-// SPDX-License-Identifier: GPL-3.0-or-later
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
-
-package com.xenoterracide.git;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.xenoterracide.git.internal.NoExecSystemReader;
-import org.junit.jupiter.api.Test;
-
-class NoExecSystemReaderTest {
-
-  @Test
-  void getEnvPathReturnsEmptyString() {
-    var reader = new NoExecSystemReader();
-    assertThat(reader.getenv("PATH")).isEmpty();
-  }
-
-  @Test
-  void getEnvOtherVariableReturnsValue() {
-    var reader = new NoExecSystemReader();
-    // Compare with actual system environment to verify pass-through behavior
-    assertThat(reader.getenv("HOME")).isEqualTo(System.getenv("HOME"));
-  }
-}
