+++ module/semver-calculator/pom.xml
@@ -19,7 +19,7 @@ SPDX-License-Identifier: MIT
     <relativePath>../../pom.xml</relativePath>
   </parent>
 
-  <artifactId>semver</artifactId>
+  <artifactId>semver-calculator</artifactId>
   <name>Semantic Versioning</name>
   <description>Semantic versioning calculation from git metadata</description>
